api/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
tortoise-orm[asyncpg]>=0.24.2
